+++ test/support/multi_id_entity.ex
@@ -1,0 +1,6 @@
+defmodule Roux.Test.MultiIdEntity do
+  @moduledoc false
+  use Roux.Entity,
+    identity: [:module_name, :name],
+    tracked: [:arity]
+end