+++ test/support/sample_entity.ex
@@ -1,0 +1,6 @@
+defmodule Roux.Test.SampleEntity do
+  @moduledoc false
+  use Roux.Entity,
+    identity: [:name],
+    tracked: [:body, :return_type]
+end